Combine ChannelQuality enum with map, tidy of comments and type hints
